+++ engine/movelist.hpp
@@ -69,16 +69,16 @@ struct MoveList {
     inline const T& operator[](size_t i) const noexcept { return *ptr(i); }
 
     inline T* begin() noexcept { return ptr(0); }
-    inline T* end() noexcept { return ptr(static_cast<size_t>(size)); }
+    inline T* end() noexcept { return ptr(size); }
     inline const T* begin() const noexcept { return ptr(0); }
-    inline const T* end() const noexcept { return ptr(static_cast<size_t>(size)); }
+    inline const T* end() const noexcept { return ptr(size); }
 
     [[nodiscard]] inline bool is_empty() const noexcept { return size == 0; }
 
     inline void clear() noexcept {
         if constexpr (!std::is_trivially_destructible_v<T>) {
             for (int i = 0; i < size; ++i) {
-                ptr(static_cast<size_t>(i))->~T();
+                ptr(i)->~T();
             }
         }
         size = 0;
@@ -94,14 +94,14 @@ struct MoveList {
         if (size <= 1) return; // nothing to sort
         
         for (int i = 1; i < size; ++i) {
-            T key = (*this)[static_cast<size_t>(i)];
+            T key = (*this)[i];
             int j = i - 1;
 
-            while (j >= 0 && ((*this)[static_cast<size_t>(j)].score < key.score )) {
-                (*this)[static_cast<size_t>(j + 1)] = (*this)[static_cast<size_t>(j)];
+            while (j >= 0 && ((*this)[j].score < key.score )) {
+                (*this)[j + 1] = (*this)[j];
                 --j;
             }
-            (*this)[static_cast<size_t>(j + 1)] = key;
+            (*this)[j + 1] = key;
         }
     }
 
@@ -112,15 +112,15 @@ struct MoveList {
         const int n = (size < LIMIT) ? size : LIMIT;
 
         for (int i = 1; i < n; ++i) {
-            T key = (*this)[static_cast<size_t>(i)];
+            T key = (*this)[i];
             int j = i - 1;
             
             // Descending order (highest score first)
-            while (j >= 0 && (*this)[static_cast<size_t>(j)].score < key.score) {
-                (*this)[static_cast<size_t>(j + 1)] = (*this)[static_cast<size_t>(j)];
+            while (j >= 0 && (*this)[j].score < key.score) {
+                (*this)[j + 1] = (*this)[j];
                 --j;
             }
-            (*this)[static_cast<size_t>(j + 1)] = key;
+            (*this)[j + 1] = key;
         }
     }
 
@@ -133,7 +133,7 @@ private:
     inline void copyFrom(const MoveList& other) noexcept(std::is_nothrow_copy_constructible_v<T>) {
         size = 0;
         for (int i = 0; i < other.size; ++i) {
-            new (&data[i]) T(*other.ptr(static_cast<size_t>(i)));
+            new (&data[i]) T(*other.ptr(i));
             ++size;
         }
     }
@@ -141,11 +141,10 @@ private:
     inline void moveFrom(MoveList&& other) noexcept(std::is_nothrow_move_constructible_v<T>) {
         size = 0;
         for (int i = 0; i < other.size; ++i) {
-            new (&data[i]) T(std::move(*other.ptr(static_cast<size_t>(i))));
+            new (&data[i]) T(std::move(*other.ptr(i)));
             ++size;
         }
         other.clear();
     }
 
 };
-
